+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## v0.3.0 (2023.03.31)
+
+### 新增
+- 增加飞书群组机器人推送  [#2](https://github.com/zema1/watchvuln/issues/2) [#8](https://github.com/zema1/watchvuln/issues/8) [#11](https://github.com/zema1/watchvuln/issues/11)
+- 增加自定义 webhook 服务的方式 [#10](https://github.com/zema1/watchvuln/pull/10),  感谢 [@lzskyline](https://github.com/lzskyline)
+- 增加漏洞等级和漏洞标签变更的支持，比如开始是低危，后面改成高危了也可以正常推送
+- 增加 `--no-filter` 选项可以禁用内置的漏洞过滤器
+- 增加 `--no-start-message` 选项可以禁用启动时的提示信息
+- 整理 README 文档，使用更清晰
+
+### 修复
+- 修复 Docker 容器的时区问题，改为 `Asia/Shanghai` [#7](https://github.com/zema1/watchvuln/issues/7)
+
 ## v0.2.0 (2023.03.27)
 
 ### 新增